+++ src/django/pfb_analysis/management/commands/add_city_fips.py
@@ -52,7 +52,7 @@ def add_city_fips(input_path, output_path, fips_csv):
             for feature in shp:
                 prop = feature.get('properties')
                 city = prop['city'].upper()
-                state_places = lookup.get(prop['state']).keys()
+                state_places = list(lookup.get(prop['state']).keys())
                 found_fips = None
                 if state_places:
                     for place in state_places:
@@ -63,13 +63,13 @@ def add_city_fips(input_path, output_path, fips_csv):
                                                                   state=prop['state'], fips=fips))
                             if found_fips:
                                 if found_fips != fips:
-                                    logger.warn("""FIPS mismatch. Expected: {fips}
+                                    logger.warning("""FIPS mismatch. Expected: {fips}
                                         Got: {found} for {place}, {state}""".format(
                                         fips=fips, found=found_fips, place=place,
                                         state=prop['state']))
                             found_fips = fips
                 if not found_fips:
-                    logger.warn('Could not find FIPS for {city}, {state}'.format(city=city,
+                    logger.warning('Could not find FIPS for {city}, {state}'.format(city=city,
                                 state=prop['state']))
                 else:
                     found += 1