+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Upcoming release]
 
+## [0.11.0] - 2019-10-31
+
+#### Changed
+- Upgraded Django app to run under Python 3
+- Upgraded Django app to Django 2.2
+- Upgraded osm2pgrouting and osm2pgsql and switched them to use the apt packages
 
 ## [0.10.0] - 2019-05-08
 
@@ -20,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support entering state/province for non-US neighborhoods
 - Default to sorting places list by name and remove order-by-state option
 - Adjust label formatting to show state/province and country
+- Upgraded analysis scripts to run under Python 3
 
 ## [0.9.2] - 2019-03-11
 
@@ -98,7 +105,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.1.0] - 2017-04-21
 
 
-[Upcoming release]: https://github.com/azavea/pfb-network-connectivity/compare/0.10.0...HEAD
+[Upcoming release]: https://github.com/azavea/pfb-network-connectivity/compare/0.11.0...HEAD
+[0.11.0]: https://github.com/azavea/pfb-network-connectivity/compare/0.10.0...0.11.0
 [0.10.0]: https://github.com/azavea/pfb-network-connectivity/compare/0.9.2...0.10.0
 [0.9.2]: https://github.com/azavea/pfb-network-connectivity/compare/0.9.1...0.9.2
 [0.9.1]: https://github.com/azavea/pfb-network-connectivity/compare/0.9.0...0.9.1